+++ src/useFieldArray.ts
@@ -28,11 +28,6 @@ export default <T = any, V extends FormValues = FormValues>(
   name: string,
   { formId, defaultValue, validate }: FieldArrayConfig<T, V> = {}
 ): FieldArrayReturn<T> => {
-  invariant(
-    !name,
-    '💡 react-cool-form > useFieldArray: Missing "name" parameter.'
-  );
-
   const methods: Methods<V> = shared.get(formId);
 
   invariant(
@@ -57,7 +52,7 @@ export default <T = any, V extends FormValues = FormValues>(
 
   const getFields = useCallback(
     (init = false): string[] => {
-      let fields = getState(`values.${name}`);
+      let fields = getState(name);
 
       if (init && isUndefined(fields)) fields = defaultValueRef.current;
 
@@ -191,7 +186,7 @@ export default <T = any, V extends FormValues = FormValues>(
         f.splice(index, 1);
         return compact(f).length ? f : [];
       };
-      const value = (getState(`values.${name}`) || [])[index];
+      const value = (getState(name) || [])[index];
 
       setState(handler);